+++ easytk/window.py
@@ -1,5 +1,5 @@
 from easytk.processes import ETKQueueManager, ETKBackgroundProcessManager
-from easytk.widgets import ETKWidgetManager
+from easytk.widgets import ETKWidgetManager, ETKFrameManager
 from easytk.state import ETKStateManager
 from tkinter import Tk, Toplevel
 
@@ -9,6 +9,7 @@ class ETKWindow:
     state: ETKStateManager
     processes: ETKBackgroundProcessManager
     queue: ETKQueueManager
+    frames: ETKFrameManager
     master: Tk | Toplevel
     def __init__(self, master: Tk|Toplevel):
         self.master = master
@@ -16,11 +17,31 @@ class ETKWindow:
         self.state = ETKStateManager(self.master)
         self.processes = ETKBackgroundProcessManager()
         self.queue = ETKQueueManager(self.master)
+        self.frames = ETKFrameManager()
+    
     
     def sub(self): return ETKWindow(Toplevel(self.master))
     
     @property
     def m(self): return self.master    
+        
+    def geometry(self, geometry_string: str):
+        """Set the window geometry (size and position).
+        
+        Args:
+            geometry_string (str): Geometry string in format 'widthxheight+x+y' or 'widthxheight'
+        """
+        self.master.geometry(geometry_string)
+        return self
+    
+    def title(self, title: str):
+        """Set the window title.
+        
+        Args:
+            title (str): The title to set for the window
+        """
+        self.master.title(title)
+        return self
         
     def run(self): 
         self.master.mainloop()